# api/app/Docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestao de Propostas API",
        "description": "API REST para gestao de propostas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080"
        }
    ],
    "paths": {
...
        "/api/v1/propostas": {
            "get": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Lista propostas",
                "operationId": "6f82ee420dfc54b6daeaba46a931b2e8",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/status"
                    },
                    {
                        "$ref": "#/components/parameters/origem"
                    },
                    {
                        "$ref": "#/components/parameters/cliente_id"
                    },
                    {
                        "$ref": "#/components/parameters/produto"
                    },
                    {
                        "$ref": "#/components/parameters/min_valor"
                    },
                    {
                        "$ref": "#/components/parameters/max_valor"
                    },
                    {
                        "$ref": "#/components/parameters/created_from"
                    },
                    {
                        "$ref": "#/components/parameters/created_to"
                    },
                    {
                        "$ref": "#/components/parameters/sort"
                    },
                    {
                        "$ref": "#/components/parameters/direction"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de propostas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PropostaList"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Cria uma proposta",
                "operationId": "64522b99bfedef89fa78a10474edb031",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/IdempotencyKey"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PropostaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proposta criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proposta"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/propostas/{id}": {
            "get": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Detalha uma proposta",
                "operationId": "8bec26a46847849e8842e487ac9e8bee",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proposta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proposta"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Exclusao logica de proposta",
                "operationId": "0a9ed1b0be7e3981c7d286f849cd65d3",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PropostaDelete"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proposta excluida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proposta"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Atualiza campos sensiveis",
                "operationId": "fc6f226c3034bc079d91cbf03a981df8",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PropostaUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proposta atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proposta"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PropostaList": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Proposta"
                        }
                    },
                    "meta": {
                        "properties": {
                            "page": {
                                "type": "integer"
                            },
                            "per_page": {
                                "type": "integer"
                            },
                            "total": {
                                "type": "integer"
                            },
                            "page_count": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "status": {
                "name": "status",
                "in": "query",
                "schema": {
                    "type": "string"
                }
            },
            "origem": {
                "name": "origem",
                "in": "query",
                "schema": {
                    "type": "string"
                }
            },
            "cliente_id": {
                "name": "cliente_id",
                "in": "query",
                "schema": {
                    "type": "integer"
                }
            },
            "produto": {
                "name": "produto",
                "in": "query",
                "schema": {
                    "type": "string"
                }
            },
            "min_valor": {
                "name": "min_valor",
                "in": "query",
                "schema": {
                    "type": "number",
                    "format": "float"
                }
            },
            "max_valor": {
                "name": "max_valor",
                "in": "query",
                "schema": {
                    "type": "number",
                    "format": "float"
                }
            },
            "created_from": {
                "name": "created_from",
                "in": "query",
                "schema": {
                    "type": "string",
                    "format": "date"
                }
            },
            "created_to": {
                "name": "created_to",
                "in": "query",
                "schema": {
                    "type": "string",
                    "format": "date"
                }
            },
            "sort": {
                "name": "sort",
                "in": "query",
                "schema": {
                    "type": "string"
                }
            },
            "direction": {
                "name": "direction",
                "in": "query",
                "schema": {
                    "type": "string"
                }
            },
            "page": {
                "name": "page",
                "in": "query",
                "schema": {
                    "type": "integer"
                }
            },
            "per_page": {
                "name": "per_page",
                "in": "query",
                "schema": {
                    "type": "integer"
                }
            }
        }
    },
...
}